LayoutTests/platform/efl/fast/events/platform-wheelevent-in-scrolling-div-expected.txt

FAIL event.wheelDeltaY should be -600. Was 0.
PASS event.wheelDeltaX is window.expectedScrollLeft*-120
PASS event.wheelDelta is window.expectedScrollLeft*-120
PASS event.wheelDeltaY is window.expectedScrollTop*-120
FAIL event.wheelDeltaX should be -360. Was 0.
PASS event.wheelDelta is window.expectedScrollTop*-120
PASS div.scrollTop is window.expectedScrollTop*window.pixelsPerWheelTick
PASS div.scrollLeft is window.expectedScrollLeft*window.pixelsPerWheelTick
